RF: Breaking down the tests and utilities to remove possible circular dependencies.
