+++ fury/actor/billboard.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 import numpy as np
 
-from fury.actor.core import Mesh
+from fury.actor import Mesh
 from fury.geometry import buffer_to_geometry
 from fury.lib import register_wgpu_render_function
 from fury.material import BillboardMaterial, validate_opacity
