+++ fury/actor/curved.py
@@ -4,8 +4,7 @@ from concurrent.futures import ThreadPoolExecutor
 
 import numpy as np
 
-from fury.actor import actor_from_primitive, create_mesh
-from fury.actor.core import Line, Mesh
+from fury.actor import Line, Mesh, actor_from_primitive, create_mesh
 from fury.geometry import buffer_to_geometry, line_buffer_separator
 from fury.lib import (
     Buffer,
