+++ fury/actor/planar.py
@@ -6,8 +6,13 @@ import logging
 import numpy as np
 from scipy.spatial.transform import Rotation as R
 
-from fury.actor import actor_from_primitive, create_image, create_point, create_text
-from fury.actor.core import Points
+from fury.actor import (
+    Points,
+    actor_from_primitive,
+    create_image,
+    create_point,
+    create_text,
+)
 from fury.geometry import (
     buffer_to_geometry,
     line_buffer_separator,
