+++ fury/actor/tests/_helpers.py
@@ -1,0 +1,97 @@
+"""Shared helpers for actor module tests."""
+
+from PIL import Image
+import numpy as np
+import numpy.testing as npt
+
+from fury import actor, window
+
+
+def random_png(width, height):
+    """Generates a random RGB PNG image.
+
+    Parameters
+    ----------
+    width : int
+        Width of the image in pixels.
+    height : int
+        Height of the image in pixels.
+
+    Returns
+    -------
+    Image
+        The generated image.
+    """
+    image = Image.new("RGB", (width, height))
+    pixels = image.load()
+
+    for x in range(width):
+        for y in range(height):
+            r = np.random.randint(0, 255)
+            g = np.random.randint(0, 255)
+            b = np.random.randint(0, 255)
+            pixels[x, y] = (r, g, b)
+
+    return image
+
+
+def validate_actors(actor_type="actor_name", prim_count=1, **kwargs):
+    scene = window.Scene()
+    typ_actor = getattr(actor, actor_type)
+    get_actor = typ_actor(**kwargs)
+    scene.add(get_actor)
+
+    centers = kwargs.get("centers", None)
+    colors = kwargs.get("colors", None)
+
+    if centers is not None:
+        npt.assert_array_equal(get_actor.local.position, centers[0])
+
+        mean_vertex = np.round(np.mean(get_actor.geometry.positions.view, axis=0))
+        npt.assert_array_almost_equal(mean_vertex, centers[0])
+
+    assert get_actor.prim_count == prim_count
+
+    if actor_type == "line":
+        return
+
+    fname = f"{actor_type}_test.png"
+    window.snapshot(scene=scene, fname=fname)
+
+    img = Image.open(fname)
+    img_array = np.array(img)
+
+    mean_r, mean_g, mean_b, _mean_a = np.mean(
+        img_array.reshape(-1, img_array.shape[2]), axis=0
+    )
+
+    assert mean_r > mean_b and mean_r > mean_g
+
+    middle_pixel = img_array[img_array.shape[0] // 2, img_array.shape[1] // 2]
+    r, g, b, a = middle_pixel
+    assert r > g and r > b
+    assert g == b
+    scene.remove(get_actor)
+
+    typ_actor_1 = getattr(actor, actor_type)
+    get_actor_1 = typ_actor_1(centers=centers, colors=colors, material="basic")
+    scene.add(get_actor_1)
+    fname_1 = f"{actor_type}_test_1.png"
+    window.snapshot(scene=scene, fname=fname_1)
+    img = Image.open(fname_1)
+    img_array = np.array(img)
+
+    mean_r, mean_g, mean_b, _mean_a = np.mean(
+        img_array.reshape(-1, img_array.shape[2]), axis=0
+    )
+
+    assert mean_r > mean_b and mean_r > mean_g
+    assert 0 < mean_r < 255
+    assert mean_g == 0 and mean_b == 0
+
+    middle_pixel = img_array[img_array.shape[0] // 2, img_array.shape[1] // 2]
+    r, g, b, a = middle_pixel
+    assert r > g and r > b
+    assert g == 0 and b == 0
+    assert r == 255
+    scene.remove(get_actor_1)
