+++ fury/actor/tests/test_bio.py
@@ -1,8 +1,7 @@
 import numpy as np
 import pytest
 
-from fury.actor import peaks_slicer, volume_slicer
-from fury.actor.core import Group
+from fury.actor import Group, peaks_slicer, volume_slicer
 from fury.lib import AffineTransform
