+++ fury/ui/tests/test_core.py
@@ -5,7 +5,7 @@ import numpy as np
 import numpy.testing as npt
 
 from fury import ui, window
-from fury.actor.core import Mesh
+from fury.actor import Mesh
 from fury.ui.helpers import Anchor
